+++ libs/application/templates/estate/src/forms/Overview/overviewConfirmAction.ts
@@ -1,0 +1,35 @@
+import {
+  buildCheckboxField,
+  getValueViaPath,
+} from '@island.is/application/core'
+import { m } from '../../lib/messages'
+import { EstateTypes, NO, YES } from '../../lib/constants'
+
+export const overviewConfirmAction = [
+  buildCheckboxField({
+    id: 'overviewConfirmAction',
+    title: '',
+    backgroundColor: 'blue',
+    defaultValue: [],
+    condition: (answers) =>
+      getValueViaPath(answers, 'selectedEstate') ===
+      EstateTypes.estateWithoutAssets,
+    options: (application) =>
+      getValueViaPath(
+        application.answers,
+        'estateWithoutAssets.estateAssetsExist',
+      ) === NO
+        ? [
+            {
+              value: YES,
+              label: m.acceptNoAssets.defaultMessage,
+            },
+          ]
+        : [
+            {
+              value: YES,
+              label: m.acceptAssets.defaultMessage,
+            },
+          ],
+  }),
+]
